backend/pyproject.toml
[project]
name = "pokeland-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Leonardo Paiva",email = "leonardo.paivx@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "flask[async] (>=3.1.0,<4.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "alembic (>=1.14.0,<2.0.0)",
    "pydantic[email] (>=2.10.5,<3.0.0)",
    "flask-pydantic (>=0.12.0,<0.13.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest-asyncio (>=0.25.2,<0.26.0)",
    "pytest-mock (>=3.14.0,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
taskipy = "^1.14.1"
ruff = "^0.9.1"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'
asyncio_mode = "auto"

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'flask --app main run --debug'
pre_test = 'task lint'
test = 'pytest -s -x --cov=backend -vv'
post_test = 'coverage html'
